% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegEqualSigma.R
\name{RegEqualSigma}
\alias{RegEqualSigma}
\title{Regression of data with Equal Uncertanties}
\usage{
RegEqualSigma(Xi, Yi, sXi, sYi, plot = F)
}
\arguments{
\item{Xi}{Numeric vector containing the x variable measured values}

\item{Yi}{Numeric vector containing the y variable measured values}

\item{sXi}{Numeric vector containing the uncertainties of the x measured values}

\item{sYi}{Numeric vector containing the uncertainties of the y measured values}

\item{plot}{Default False. Will return plots if True of: i) iterativly computed b values against iteration number,
ii) Y vs X plot with the OLS regression in blue and York linear regression in red,
iii) Chi-Squared density function for the corresponding DF, with the S statistic
as a vertical red line.

The functions returns:}
}
\value{
York slope and York intercept of York regression,Xi-xi,Yi-yi,sqrt(Sb2),sqrt(Sa2),S,Q,length(Xi)-2, bVect

Expected xi and Expected yi - Expected x and x values,

ResXi,ResYi - Residuals of all Xi and Yi measurement respect to the corresponding expected xi and yi,

Rsquared - Coefficient of determinaton on the computed linear model,

MAE_X,MAE_y,MAEavg - Mean Absolute Error on X and Y directions and Average Mean Absolute Error,

sqrt(Sb2),sqrt(Sa2) - Uncertainty on slope and intercept,

S - the minimized sum of squared weighted residuals,

Q - probability of obtaining the computed S value or higher for the relative Chi-Squared density function for the corresponding DF,

length(Xi)-2 - the DF degrees of freedom,
}
\description{
This code implements a function to perform linear regression on x and y data which as equal measurement
uncertainties. The function depend on 'ggplot2' for plotting.
The function RegEqualSigma() require the following parameters.
}
