% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YorkRegression.R
\name{YorkRegression}
\alias{YorkRegression}
\title{York linear regression function}
\usage{
YorkRegression(Xi, Yi, sXi, sYi, iter = 100, plot = F)
}
\arguments{
\item{Xi}{Numeric vector containing the x variable measured values}

\item{Yi}{Numeric vector containing the y variable measured values}

\item{sXi}{Numeric vector containing the uncertainties of the x measured values}

\item{sYi}{Numeric vector containing the uncertainties of the y measured values}

\item{iter}{Number of iterations to be performed, default is 100}

\item{plot}{Default False. Will return plots if True of: i) iterativly computed b values against iteration number,
ii) Y vs X plot with the OLS regression in blue and York linear regression in red,
iii) Chi-Squared density function for the corresponding DF, with the S statistic
as a vertical red line.

The functions returns:}
}
\value{
York slope and York intercept of York regression,Xi-xi,Yi-yi,sqrt(Sb2),sqrt(Sa2),S,Q,length(Xi)-2, bVect

Expected xi and Expected yi - Expected x and x values,

Xi-xi,Yi-yi - Residuals of all Xi and Yi measurement respect to the corresponding expected xi and yi,

sqrt(Sb2),sqrt(Sa2) - Uncertainty on slope and intercept,

Rsquared - Coefficient of determinaton on the computed linear model,

MAE_X,MAE_y,MAEdiag - Mean Absolute Error on X and Y directions and Diagonal Mean Absolute Error,

S - the minimized sum of squared weighted residuals,

Q - probability of obtaining the computed S value or higher for the relative Chi-Squared density function for the corresponding DF,

length(Xi)-2 - the DF degrees of freedom,

bVect - Iterations of b value computation.
}
\description{
This code implements a function to perform York linear regression which accounts for
uncertainties in both y and y variable, with uncertainties also potentially correlated.
The function depend on 'ggplot2' for plotting.
The function YorkRegression() require the following parameters.
}
